Move stash logic to QueueDB
