+++ mobile/src/SyncFacade.js
@@ -1,7 +1,7 @@
 import {appIndexDb} from "./AppDatabase";
 import {ApiServices} from "./Services/ApiServices";
-import {getSelectedProgram, saveSelectedProgram} from "./components/ProgramSelection";
 import {programDb} from "./Services/ProgramDB";
+import {queueDb} from "./Services/QueueDB";
 
 const LAST_SYNC_KEY = "lastSyncedDate";
 
@@ -28,7 +28,7 @@ export class SyncFacade {
         const vaccinators = await ApiServices.fetchVaccinators();
         await appIndexDb.saveVaccinators(vaccinators);
 
-        await appIndexDb.popData()
+        await queueDb.popData()
     }
 
     static async push() {
